capstone/requirements.txt
jayson
lxml==3.3.5
newspaper
nltk==2.0.5
numpy
pandas
pika
Pillow==2.5.1
python-dateutil
python-jsonrpc
pymongo
PyYAML==3.11
redis
requests
six
scipy
sklearn
tldextract==1.5.1
watchdog